pyproject.toml
[project]
name = "tool-google-calendar"
version = "0.1.0"
description = "Ahnopologetic tools - Google Calendar"
authors = [
    {name = "ahnopologetic", email = "stahn1995@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client",
    "google-auth-oauthlib",
    "pyyaml (>=6.0.2,<7.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^3.0"
ruff = "^0.9.2"
openai = "^1.59.9"
pytest-mock = "^3.14.0"
mypy = "^1.14.1"
google-api-python-client-stubs = "^1.29.0"
pandas-stubs = "^2.2.3.241126"
types-pyyaml = "^6.0.12.20241230"
types-pygments = "^2.19.0.20250107"
types-cachetools = "^5.5.0.20240820"
types-colorama = "^0.4.15.20240311"
types-python-dateutil = "^2.9.0.20241206"
types-setuptools = "^75.8.0.20250110"

[tool.pytest.ini_options]
markers = [
    "local_only: mark a test as local only",
]

